DynamicChangLog: fix the UI components not showing correctly
I've about cleared out my TODO lists but previous commit was
ripe with derps that needed to be addressed right away.

The only thing that is still not working is the sorting
it seems to find the date and put them in order but the
screen never updates to new order (weird). If we can't
fix it I may be forced to remove the sorting, even though
that is the reason I wanted to work on this project.

Possible reasons the UI isn't updating, anyone?

still TODO: move UI tasks to Runnables
          : while reading packages update mCategory title
            (now it gets legacy title from packages list)
          : add Exception catching to packages listings,
            we handle null author and null committer while
            scrolling through for recent commits but not
            when the screen comes from packages list